+++ src/game/gold/mod.rs
@@ -1,5 +1,4 @@
 use bevy::prelude::*;
-use crate::player::*;
 use crate::common::components::BoundingBox;
 use crate::player::components::Player;
