--- src/window_icon.rs
@@ -1,24 +1,0 @@
-use bevy::winit::WinitWindows;
-use bevy::{prelude::*, window::PrimaryWindow};
-use winit::window::Icon;
-
-
-// https://stackoverflow.com/questions/74586997/how-to-add-a-window-icon-in-bevy
-pub fn set_window_icon(
-    main_window: Query<Entity, With<PrimaryWindow>>,
-    windows: NonSend<WinitWindows>,
-) {
-    let Some(primary) = windows.get_window(main_window.single()) else {return};
-
-    let (icon_rgba, icon_width, icon_height) = {
-        let image = image::open("assets/window_icon.png")
-            .expect("Failed to open icon path")
-            .into_rgba8();
-        let (width, height) = image.dimensions();
-        let rgba = image.into_raw();
-        (rgba, width, height)
-    };
-
-    let icon = Icon::from_rgba(icon_rgba, icon_width, icon_height).unwrap();
-    primary.set_window_icon(Some(icon));
-}
